NURO-57 refactor foldering
