+++ Nuro/Features/Parent/Routine/Controllers/ParentRoutineViewController.swift
@@ -1,0 +1,23 @@
+//
+//  ParentRoutineViewController.swift
+//  Nuro
+//
+//  Created by Samuel Dennis on 05/10/22.
+//
+
+import UIKit
+
+class ParentRoutineViewController: UIViewController {
+
+    let parentRoutineView = ParentRoutineView()
+    let viewModel = ParentRoutineViewModel()
+    
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        parentRoutineView.setup(vc: self)
+    }
+    
+    override func loadView() {
+        self.view = parentRoutineView
+    }
+}
