+++ Nuro/Features/Parent/Routine/ParentRoutineDelegate.swift
@@ -1,0 +1,12 @@
+//
+//  ParentRoutineDelegate.swift
+//  Nuro
+//
+//  Created by Samuel Dennis on 05/10/22.
+//
+
+import Foundation
+
+protocol ParentRoutineDelegate {
+    func printText()
+}
