+++ Nuro/Features/Parent/Routine/ViewModels/ParentRoutineViewModel.swift
@@ -6,3 +6,11 @@
 //
 
 import Foundation
+
+class ParentRoutineViewModel {
+    
+    func printText() {
+        print("Segmented Selected")
+    }
+    
+}
